+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- WrapperWidgetUI: contentProps property to support scrolling [#769](https://github.com/CartoDB/carto-react/pull/769)
 
 ## 2.2