+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Not released
 
+## 2.2
+
+### 2.2.7 (2023-09-13)
+
 - SelectField: Remove empty space at the beginning [#772](https://github.com/CartoDB/carto-react/pull/772)
 - WrapperWidgetUI: contentProps property to support scrolling [#769](https://github.com/CartoDB/carto-react/pull/769)
-
-## 2.2
 
 ### 2.2.6 (2023-09-07)